Change all diff int types to int64_t

+++ GraphTree/DateTime/include/DateTime.h
@@ -23,35 +23,35 @@ namespace DT {
     class DateTimeDelta;
 
     // Return true if date is valid, return false otherwise.
-    bool validateDate(std::uint32_t d, std::uint32_t m, std::uint32_t y);
+    bool validateDate(std::int64_t d, std::int64_t m, std::int64_t y);
 
     // Returns an integer that represents week day (Mon=1, Tue=2...) if date is invalid return -1.
-    std::int32_t getDayOfWeek(std::uint32_t d, std::uint32_t m, std::uint32_t y);
+    std::int64_t getDayOfWeek(std::int64_t d, std::int64_t m, std::int64_t y);
 
     // Returns day name from integer 1..7
-    std::string toDayOfWeek(std::int32_t day);
+    std::string toDayOfWeek(std::int64_t day);
 
     // Return true if year is a leap year, false otherwise.
-    bool isLeapYear(std::uint32_t year);
+    bool isLeapYear(std::int64_t year);
 
     // Return total number of days.
-    std::int32_t numberOfDays(const DateTime &dt);
+    std::int64_t numberOfDays(const DateTime &dt);
 
-    Date daysToDate(std::uint32_t daysCount);
+    Date daysToDate(std::int64_t daysCount);
 
 
 
     class Date {
     public:
-        explicit Date(std::uint32_t d, std::uint32_t m, std::uint32_t y);
+        explicit Date(std::int64_t d, std::int64_t m, std::int64_t y);
 
         ~Date() = default;
 
-        std::uint32_t Month() const { return month; }
+        std::int64_t Month() const { return month; }
 
-        std::uint32_t Day() const { return day; }
+        std::int64_t Day() const { return day; }
 
-        std::uint32_t Year() const { return year; }
+        std::int64_t Year() const { return year; }
 
         void print();
 
@@ -66,14 +66,14 @@ namespace DT {
         bool const operator>(const Date &rsv);
 
     private:
-        std::uint32_t month;
-        std::uint32_t day;
-        std::uint32_t year;
+        std::int64_t month;
+        std::int64_t day;
+        std::int64_t year;
     };
 
     class Time {
     public:
-        explicit Time(std::uint32_t h, std::uint32_t m, std::uint32_t s);
+        explicit Time(std::int64_t h, std::int64_t m, std::int64_t s);
 
         ~Time() = default;
 
@@ -87,17 +87,17 @@ namespace DT {
 
         bool const operator>(const Time &rsv);
 
-        std::uint32_t Hour() const { return hour; }
+        std::int64_t Hour() const { return hour; }
 
-        std::uint32_t Minute() const { return minute; }
+        std::int64_t Minute() const { return minute; }
 
-        std::uint32_t Seconds() const { return seconds; }
+        std::int64_t Seconds() const { return seconds; }
 
 
     private:
-        std::uint32_t hour, minute, seconds;
+        std::int64_t hour, minute, seconds;
 
-        bool checkInRange(std::uint32_t val, std::uint32_t a, std::uint32_t b) { return (val >= a && val <= b); }
+        bool checkInRange(std::int64_t val, std::int64_t a, std::int64_t b) { return (val >= a && val <= b); }
 
         double getFrac();
     };
@@ -111,19 +111,19 @@ namespace DT {
 
         ~DateTime() = default;
 
-        std::int32_t dayOfWeek() const;
+        std::int64_t dayOfWeek() const;
 
-        std::uint32_t Month() const { return date.Month(); }
+        std::int64_t Month() const { return date.Month(); }
 
-        std::uint32_t Day() const { return date.Day(); }
+        std::int64_t Day() const { return date.Day(); }
 
-        std::uint32_t Year() const { return date.Year(); }
+        std::int64_t Year() const { return date.Year(); }
 
-        std::uint32_t Hour() const { return time.Hour(); }
+        std::int64_t Hour() const { return time.Hour(); }
 
-        std::uint32_t Minute() const { return time.Minute(); }
+        std::int64_t Minute() const { return time.Minute(); }
 
-        std::uint32_t Seconds() const { return time.Seconds(); }
+        std::int64_t Seconds() const { return time.Seconds(); }
 
         void print() const;
