+++ GraphTree/DateTime/src/DateTime.cpp
@@ -10,7 +10,7 @@
 
 namespace DT {
 
-    Time::Time(const std::uint32_t h, const std::uint32_t m, const std::uint32_t s) {
+    Time::Time(const std::int64_t h, const std::int64_t m, const std::int64_t s) {
         try {
             if (checkInRange(h, 0, 23))
                 hour = h;
@@ -81,7 +81,7 @@ namespace DT {
     }
 
 
-    bool validateDate(const std::uint32_t d, const std::uint32_t m, const std::uint32_t y) {
+    bool validateDate(const std::int64_t d, const std::int64_t m, const std::int64_t y) {
         if (y >= 0 && y <= 9999) {
             //check month
             if (m >= 1 && m <= 12) {
@@ -114,21 +114,21 @@ namespace DT {
 
     }
 
-    std::int32_t getDayOfWeek(std::uint32_t d, std::uint32_t m, std::uint32_t y) {
+    std::int64_t getDayOfWeek(std::int64_t d, std::int64_t m, std::int64_t y) {
         if (!validateDate(d, m, y)) {
             return -1;
         }
         // 1990, Michael Keith and Tom Craver expression.
-        std::int32_t day = (d += m < 3 ? y-- : y - 2, 23 * m / 9 + d + 4 + y / 4 - y / 100 + y / 400) % 7;
+        std::int64_t day = (d += m < 3 ? y-- : y - 2, 23 * m / 9 + d + 4 + y / 4 - y / 100 + y / 400) % 7;
         // To make Sun week day number 7 not 0.
         return (day == 0 ? 7 : day);
     }
 
-    bool checkInRange(const std::uint32_t val, const std::uint32_t a, const std::uint32_t b) {
+    bool checkInRange(const std::int64_t val, const std::int64_t a, const std::int64_t b) {
         return (val >= a && val <= b);
     }
 
-    Date::Date(const std::uint32_t d, const std::uint32_t m, const std::uint32_t y) {
+    Date::Date(const std::int64_t d, const std::int64_t m, const std::int64_t y) {
         if (validateDate(d, m, y)) {
             month = m;
             day = d;
@@ -194,7 +194,7 @@ namespace DT {
         cout << time.Seconds() << '\n';
     }
 
-    std::int32_t DateTime::dayOfWeek() const {
+    std::int64_t DateTime::dayOfWeek() const {
         return DT::getDayOfWeek(date.Day(), date.Month(), date.Year());
     }
 
@@ -238,16 +238,16 @@ namespace DT {
     }
 
     DateTime DateTime::operator+(const DateTimeDelta &rsv) {
-        std::uint32_t newDay = date.Day();
-        std::uint32_t newMonth = date.Month();
-        std::uint32_t newYear = date.Year();
+        std::int64_t newDay = date.Day();
+        std::int64_t newMonth = date.Month();
+        std::int64_t newYear = date.Year();
 
-        std::uint32_t newHour = time.Hour();
-        std::uint32_t newMinute = time.Minute();
-        std::uint32_t newSeconds = time.Seconds();
+        std::int64_t newHour = time.Hour();
+        std::int64_t newMinute = time.Minute();
+        std::int64_t newSeconds = time.Seconds();
 
 
-        std::uint32_t prevDays = numberOfDays(*this);
+        std::int64_t prevDays = numberOfDays(*this);
 
         Date newDate(newDay, newMonth, newYear);
         Time newTime(newHour, newMinute, newSeconds);
@@ -258,7 +258,7 @@ namespace DT {
 
 
 
-    std::string toDayOfWeek(const std::int32_t day) {
+    std::string toDayOfWeek(const std::int64_t day) {
         switch (day) {
             case -1:
                 return "INVALID";
@@ -281,7 +281,7 @@ namespace DT {
         }
     }
 
-    bool isLeapYear(std::uint32_t year) {
+    bool isLeapYear(std::int64_t year) {
         if (year % 4 == 0)
             return true;
         else if ((year % 100 == 0) && (year % 400 == 0))
@@ -290,16 +290,16 @@ namespace DT {
             return false;
     }
 
-    std::int32_t numberOfDays(const DateTime &dt) {       /* convert date to day number */
-        std::int32_t y, m;
+    std::int64_t numberOfDays(const DateTime &dt) {       /* convert date to day number */
+        std::int64_t y, m;
 
         m = (dt.Month() + 9) % 12;                /* mar=0, feb=11 */
         y = dt.Year() - m / 10;                     /* if Jan/Feb, year-- */
         return y * 365 + y / 4 - y / 100 + y / 400 + (m * 306 + 5) / 10 + (dt.Day() - 1);
     }
 
-    Date daysToDate(std::uint32_t daysCount) { /* convert day number to y,m,d format */
-        std::uint32_t y, ddd, mi, mm, dd;
+    Date daysToDate(std::int64_t daysCount) { /* convert day number to y,m,d format */
+        std::int64_t y, ddd, mi, mm, dd;
         y = (10000 * daysCount + 14780) / 3652425;
         ddd = daysCount - (365 * y + y / 4 - y / 100 + y / 400);
         if (ddd < 0) {
